fix: 프로덕션 강화 — BigDecimal 마이그레이션, Rate Limiting, MongoDB 인덱스, 보안 수정
- 모든 금융 필드 double → BigDecimal 전환 (부동소수점 오차 제거)
- IP 기반 Rate Limiting 필터 추가 (일반 120/min, 거래 30/min)
- MongoDB 복합 인덱스 추가 (Order, TradeRecord, QuantProduct 등)
- WebSocket CORS trim 처리, console.error 제거, spring-dotenv 적용

+++ backend/src/main/java/com/project/whalearc/trade/dto/OrderRequest.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class OrderRequest {
@@ -25,9 +27,9 @@ public class OrderRequest {
     private String orderMethod;
 
     @DecimalMin(value = "0.00000001", message = "수량은 0보다 커야 합니다.")
-    private double quantity;
+    private BigDecimal quantity;
 
-    private Double price; // 지정가일 때만
+    private BigDecimal price; // 지정가일 때만
 
     private String assetType; // "STOCK" or "CRYPTO" (기본: CRYPTO)
 }
